# docs/swagger.yml
openapi: 3.0.0
info:
  title: API Disruptive
  description: API Disruptive
  version: 1.0.0
servers:
  - url: http://localhost:3030/api/v1
    description: Local development server
paths:
  /user:
    post:
      summary: Create a new user
      description: Create a new user
      tags:
        - User
      operationId: createUser
      requestBody:
        description: User to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/responses/User'
        '400':
          description: Wrong Data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "WRONG_DATA"
                  description:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: "Validation failed"
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "NOT_ENOUGH_PERMISSIONS"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "INTERNAL_SERVER_ERROR"
                  description:
                    type: string
                    example: "Internal server error"
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      scheme: bearer
      bearerFormat: JWT
      in: header
      name: Authorization
  schemas:
    User:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        role:
          type: string
          enum:
            - reader
            - creator
      example:
        email: "some@email.com"
        password: "somePassword1@"
        name: "My Name"
        role: "creator"
...
  responses:
    User:
      type: object
      properties:
        id:
          type: number
          example: 1
        email:
          type: string
          example: "some@email.com"
        name:
          type: string
          example: "My Name"
        role:
          type: string
          example: "creator"
        user_name:
          type: string
          example: "myname00"